requirements.txt
astral==3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
crontab==1.0.1
flup-py3==1.0.3
gpiozero==2.0
idna==3.7
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
smbus==1.1.post2
swagger-client @ git+https://github.com/weatherapicom/python.git@6e3d4ee6160afdc66412417242364162935d5971
urllib3==2.2.1
w1thermsensor==2.3.0
